fix: owner only can delete account

+++ app/api/users/[id]/route.ts
@@ -3,18 +3,21 @@ import User from "@/models/user.model";
 import { Types } from "mongoose";
 import validateBody from "@/lib/validateBody";
 import { UserSchema } from "@/lib/Schema/UserSchema";
+import dbConnect from "@/lib/dbConnect";
+import { auth } from "@/auth";
 
 //get user by id
 export async function GET(
 	request: Request,
-	{ params }: { params: Promise<{ id: string }> }
+	{ params }: { params: Promise<{ id: string }> },
 ) {
 	try {
+		await dbConnect();
 		const { id } = await params;
 
 		if (!Types.ObjectId.isValid(id)) throw new Error("Invalid User ID");
 
-		const user = await User.findById(id);
+		const user = await User.findById(id).select("-password");
 
 		if (!user) throw new Error("User not found");
 
@@ -27,12 +30,19 @@ export async function GET(
 //delete user by id
 export async function DELETE(
 	request: Request,
-	{ params }: { params: Promise<{ id: string }> }
+	{ params }: { params: Promise<{ id: string }> },
 ) {
 	try {
 		const { id } = await params;
-
 		if (!Types.ObjectId.isValid(id)) throw new Error("Invalid User ID");
+		const auth_session = await auth();
+		const userId = auth_session?.user?.id;
+		if (!auth_session || !auth_session.user) {
+			throw new Error("Unauthorized: Please login first");
+		}
+		if (userId !== id) {
+			throw new Error("Forbidden: You can only delete your own account");
+		}
 
 		const user = await User.findByIdAndDelete(id);
 
@@ -45,7 +55,7 @@ export async function DELETE(
 }
 export async function PUT(
 	request: Request,
-	{ params }: { params: Promise<{ id: string }> }
+	{ params }: { params: Promise<{ id: string }> },
 ) {
 	try {
 		const { id } = await params;
@@ -53,10 +63,19 @@ export async function PUT(
 
 		if (!Types.ObjectId.isValid(id)) throw new Error("Invalid User ID");
 
+		const auth_session = await auth();
+		const userId = auth_session?.user?.id;
+		if (!auth_session || !auth_session.user) {
+			throw new Error("Unauthorized: Please login first");
+		}
+		if (userId !== id) {
+			throw new Error("Forbidden: You can only delete your own account");
+		}
+
 		const validatedData = validateBody(body, UserSchema, true);
 		const user = await User.findByIdAndUpdate(id, validatedData.data, {
 			new: true,
-		});
+		}).select("-password");
 
 		if (!user) throw new Error("User not found");
